refactor(training): migrate mlflow utils to anemoi-utils (#397)
## Description
This is the follow up PR to
https://github.com/ecmwf/anemoi-utils/pull/174 necessary to remove the
duplicated code and instead import the functionalities migrated to
anemoi-utils

## What problem does this change solve?
See original PR

## What issue or task does this change relate to?
NA

- [x] **This PR can only be merged once the corresponding PR in
anemoi-utils has been merged and released!**

***As a contributor to the Anemoi framework, please ensure that your
changes include unit tests, updates to any affected dependencies and
documentation, and have been tested in a parallel setting (i.e., with
multiple GPUs). As a reviewer, you are also responsible for verifying
these aspects and requesting changes if they are not adequately
addressed. For guidelines about those please refer to
https://anemoi.readthedocs.io/en/latest/***

By opening this pull request, I affirm that all authors agree to the
[Contributor License
Agreement.](https://github.com/ecmwf/codex/blob/main/Legal/contributor_license_agreement.md)
